# packages/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationBackupRestoreResult": {
        "properties": {
          "imported_benchmark_reports": {
            "minimum": 0.0,
            "title": "Imported Benchmark Reports",
            "type": "integer"
          },
          "imported_jobs": {
            "minimum": 0.0,
            "title": "Imported Jobs",
            "type": "integer"
          },
          "reused_benchmark_reports": {
            "minimum": 0.0,
            "title": "Reused Benchmark Reports",
            "type": "integer"
          },
          "reused_jobs": {
            "minimum": 0.0,
            "title": "Reused Jobs",
            "type": "integer"
          },
          "total_benchmark_reports": {
            "minimum": 0.0,
            "title": "Total Benchmark Reports",
            "type": "integer"
          },
          "total_jobs": {
            "minimum": 0.0,
            "title": "Total Jobs",
            "type": "integer"
          }
        },
        "required": [
          "imported_jobs",
          "reused_jobs",
          "imported_benchmark_reports",
          "reused_benchmark_reports",
          "total_jobs",
          "total_benchmark_reports"
        ],
        "title": "ApplicationBackupRestoreResult",
        "type": "object"
      },
...
      "BenchmarkDatasetImportReceipt": {
        "properties": {
          "archive_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Archive Sha256",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_status": {
            "anyOf": [
              {
                "maximum": 599.0,
                "minimum": 400.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Status"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Request Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkDatasetImportResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": ["pending", "completed", "failed"],
            "title": "Status",
            "type": "string"
          }
        },
        "required": ["request_id", "archive_sha256", "status"],
        "title": "BenchmarkDatasetImportReceipt",
        "type": "object"
      },
      "BenchmarkDatasetImportResult": {
        "properties": {
          "imported_cases": {
            "minimum": 0.0,
            "title": "Imported Cases",
            "type": "integer"
          },
          "included_cases": {
            "minimum": 0.0,
            "title": "Included Cases",
            "type": "integer"
          },
          "included_cases_by_layout": {
            "anyOf": [
              {
                "additionalProperties": {
                  "minimum": 0.0,
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Included Cases By Layout"
          },
          "job_ids": {
            "items": {
              "type": "string"
            },
            "title": "Job Ids",
            "type": "array"
          },
          "reused_cases": {
            "minimum": 0.0,
            "title": "Reused Cases",
            "type": "integer"
          }
        },
        "required": ["imported_cases", "reused_cases", "included_cases"],
        "title": "BenchmarkDatasetImportResult",
        "type": "object"
      },
...
      "BenchmarkOverview": {
        "properties": {
          "corpus_fingerprint": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corpus Fingerprint"
          },
          "default_layout_profile": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9_]+$",
            "title": "Default Layout Profile",
            "type": "string"
          },
          "included_cases": {
            "minimum": 0.0,
            "title": "Included Cases",
            "type": "integer"
          },
          "included_cases_by_layout": {
            "additionalProperties": {
              "minimum": 0.0,
              "type": "integer"
            },
            "title": "Included Cases By Layout",
            "type": "object"
          },
          "latest_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "parser_pipelines": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkParserPipelineSummary"
            },
            "title": "Parser Pipelines",
            "type": "array"
          },
          "recent_reports": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkReportSummary"
            },
            "title": "Recent Reports",
            "type": "array"
          }
        },
        "required": [
          "included_cases",
          "included_cases_by_layout",
          "default_layout_profile"
        ],
        "title": "BenchmarkOverview",
        "type": "object"
      },
      "BenchmarkParserPipelineSummary": {
        "properties": {
          "latest_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkReportSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "layout_profile": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9_]+$",
            "title": "Layout Profile",
            "type": "string"
          },
          "parser": {
            "$ref": "#/components/schemas/PipelineOption"
          },
          "previous_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkReportSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["parser", "layout_profile"],
        "title": "BenchmarkParserPipelineSummary",
        "type": "object"
      },
...
      "BenchmarkReport": {
        "properties": {
          "accuracy": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Accuracy",
            "type": "number"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkCaseResult"
            },
            "title": "Cases",
            "type": "array"
          },
          "corpus_fingerprint": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corpus Fingerprint"
          },
          "correct_fields": {
            "minimum": 0.0,
            "title": "Correct Fields",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evaluated_fields": {
            "minimum": 0.0,
            "title": "Evaluated Fields",
            "type": "integer"
          },
          "failed_cases": {
            "minimum": 0.0,
            "title": "Failed Cases",
            "type": "integer"
          },
          "field_metrics": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkFieldMetric"
            },
            "title": "Field Metrics",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "layout_profile": {
            "title": "Layout Profile",
            "type": "string"
          },
          "parser_provider": {
            "title": "Parser Provider",
            "type": "string"
          },
          "successful_cases": {
            "minimum": 0.0,
            "title": "Successful Cases",
            "type": "integer"
          },
          "total_cases": {
            "minimum": 0.0,
            "title": "Total Cases",
            "type": "integer"
          }
        },
        "required": [
          "parser_provider",
          "layout_profile",
          "total_cases",
          "successful_cases",
          "failed_cases",
          "correct_fields",
          "evaluated_fields",
          "accuracy"
        ],
        "title": "BenchmarkReport",
        "type": "object"
      },
...
      "BenchmarkRunRequest": {
        "properties": {
          "parser_layout_profile": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9_]+$",
            "title": "Parser Layout Profile",
            "type": "string"
          },
          "parser_provider": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9_]+$",
            "title": "Parser Provider",
            "type": "string"
          }
        },
        "required": ["parser_provider", "parser_layout_profile"],
        "title": "BenchmarkRunRequest",
        "type": "object"
      },
      "BenchmarkSelectionRequest": {
        "properties": {
          "included": {
            "title": "Included",
            "type": "boolean"
          }
        },
        "required": ["included"],
        "title": "BenchmarkSelectionRequest",
        "type": "object"
      },
      "Body_admin_ocr_backups_restore": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": ["file"],
        "title": "Body_admin_ocr_backups_restore",
        "type": "object"
      },
      "Body_admin_ocr_benchmarks_import": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": ["file"],
        "title": "Body_admin_ocr_benchmarks_import",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Poker Training Analyzer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/ocr/backups/export": {
      "get": {
        "operationId": "admin_ocr_backups_export",
        "responses": {
          "200": {
            "content": {
              "application/zip": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Application Backup"
      }
    },
    "/api/admin/ocr/backups/restore": {
      "post": {
        "operationId": "admin_ocr_backups_restore",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_admin_ocr_backups_restore"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationBackupRestoreResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Backup"
      }
    },
    "/api/admin/ocr/benchmarks": {
      "get": {
        "operationId": "admin_ocr_benchmarks_get",
        "parameters": [
          {
            "in": "query",
            "name": "parser_provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[a-z0-9_]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parser Provider"
            }
          },
          {
            "in": "query",
            "name": "parser_layout_profile",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[a-z0-9_]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parser Layout Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Benchmark Overview"
      }
    },
    "/api/admin/ocr/benchmarks/export": {
      "get": {
        "operationId": "admin_ocr_benchmarks_export",
        "parameters": [
          {
            "in": "query",
            "name": "parser_provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[a-z0-9_]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parser Provider"
            }
          },
          {
            "in": "query",
            "name": "parser_layout_profile",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[a-z0-9_]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parser Layout Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Benchmark Dataset"
      }
    },
    "/api/admin/ocr/benchmarks/import": {
      "post": {
        "operationId": "admin_ocr_benchmarks_import",
        "parameters": [
          {
            "in": "header",
            "name": "X-Benchmark-Import-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Benchmark-Import-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_admin_ocr_benchmarks_import"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkDatasetImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Benchmark Dataset"
      }
    },
    "/api/admin/ocr/benchmarks/imports/{request_id}": {
      "get": {
        "operationId": "admin_ocr_benchmark_import_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkDatasetImportReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Benchmark Dataset Import"
      }
    },
    "/api/admin/ocr/benchmarks/run": {
      "post": {
        "operationId": "admin_ocr_benchmarks_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BenchmarkRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Benchmark Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Parser Benchmark"
      }
    },
    "/api/admin/ocr/benchmarks/{report_id}": {
      "get": {
        "operationId": "admin_ocr_benchmark_report_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Benchmark Report"
      }
    },
...
    "/api/admin/ocr/jobs/{job_id}/benchmark": {
      "put": {
        "operationId": "admin_ocr_job_benchmark_update",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Benchmark Inclusion"
      }
    },
...
  }
}